Remove assumption of single associated PHY
This removes the assumption that there will be a single PHY associated
with the MCU's ethernet interface. This allows for the case where an MCU
participates in an Ethernet network by interfacing directly with an
Ethernet switch IC via RMII. It also allows the MCU to access any of the
1 or more PHYs on the switch by specifying the address of the PHY when
retrieving the driver.